+++ src/main/java/ir/tic/clouddc/event/EventRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
     List<Event> findAllByActive(boolean active);
     List<Event> findAllByCategory(EventCategory eventCategory);
     @Query("SELECT COUNT(e) FROM Event e WHERE e.active = :active")
